WIP merge extent gen into DB Store
