+++ integration_tests/conftest.py
@@ -5,8 +5,6 @@ from cubedash import logs
 from cubedash.summary import FileSummaryStore, SummaryStore
 from cubedash.summary._stores import PgSummaryStore
 from datacube.index import Index
-from datacube.scripts.dataset import create_dataset, load_rules_from_types
-from datacube.utils import read_documents
 from digitalearthau.testing import factories
 from typing import Type
